Moved checkExisting to its own hook
